style: 🎨 style filter chip and move up
